♻️ Refactor Join page to server component
